+++ sdk/include/bda_types.h
@@ -11,6 +11,7 @@
 typedef unsigned char u8;
 typedef unsigned short u16;
 typedef unsigned int u32;
+typedef unsigned long long u64;
 typedef signed short s16;
 typedef int s32;
 typedef unsigned int bda_size_t;